Setup authorisations for buildings

+++ backend/building/views.py
@@ -1,13 +1,14 @@
-from rest_framework import permissions
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
+from authorisation.permissions import OwnerOfBuilding, IsAdmin, IsSuperStudent, ReadOnlyStudent, IsSyndic
 from base.models import Building
 from base.serializers import BuildingSerializer
 from util.request_response_util import *
 
 
 class DefaultBuilding(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent]
 
     def post(self, request):
         """
@@ -32,7 +33,9 @@ class DefaultBuilding(APIView):
 
 
 class BuildingIndividualView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated,
+                          IsAdmin | IsSuperStudent | ReadOnlyStudent | OwnerOfBuilding
+                          ]
 
     def get(self, request, building_id):
         """
@@ -44,6 +47,7 @@ class BuildingIndividualView(APIView):
             return bad_request(object_name="Building")
 
         building_instance = building_instance[0]
+        self.check_object_permissions(request, building_instance)
         serializer = BuildingSerializer(building_instance)
         return get_succes(serializer)
 
@@ -55,6 +59,8 @@ class BuildingIndividualView(APIView):
         if not building_instance:
             return bad_request(object_name="building")
         building_instance = building_instance[0]
+
+        self.check_object_permissions(request, building_instance)
 
         building_instance.delete()
         return delete_succes()
@@ -68,6 +74,7 @@ class BuildingIndividualView(APIView):
             return bad_request(object_name="building")
 
         building_instance = building_instance[0]
+        self.check_object_permissions(request, building_instance)
         data = request_to_dict(request.data)
 
         if "syndic" in data.keys():
@@ -84,6 +91,7 @@ class BuildingIndividualView(APIView):
 
 
 class AllBuildingsView(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent]
 
     def get(self, request):
         """
@@ -96,12 +104,16 @@ class AllBuildingsView(APIView):
 
 
 class BuildingOwnerView(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent | OwnerOfBuilding]
 
     def get(self, request, owner_id):
         """
         Get all buildings owned by syndic with given id
         """
         building_instance = Building.objects.filter(syndic=owner_id)
+
+        for b in building_instance:
+            self.check_object_permissions(request, b)
 
         if not building_instance:
             return bad_request_relation("building", "syndic")
